requirements.txt
fastapi==0.114.2
flake8==7.1.1
gpustat==1.1.1
httpx==0.27.0
kaggle==1.6.17
mlflow==2.15.0
mlflow_skinny==2.15.0
numpy<2.0.0
openpyxl==3.1.5
pandas==2.2.2
Pillow==10.4.0
psutil==6.0.0
pydantic==2.9.1
PyGithub==2.4.0
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
Requests==2.32.3
schedule==1.2.2
scikit_learn==1.5.2
streamlit==1.22.0
tensorflow==2.10.0
tensorflow_gpu==2.10.0
tqdm==4.66.4
uvicorn==0.30.6
